Get bounding box of a track.

Get bounding box of a track.

bbox(x, ...)

# S3 method for class 'track_xy'
bbox(x, spatial = TRUE, buffer = NULL, ...)

# S3 method for class 'steps_xy'
bbox(x, spatial = TRUE, buffer = NULL, ...)

Arguments

x

[track_xy, track_xyt]
A track created with make_track.

...

Further arguments, none implemented.

spatial

[logical(1)=TRUE]
Whether or not to return an object of class sf-Polygon-object or not.

buffer

[numeric(0)=NULL]{NULL, >0}
An optional buffer of the bounding box.

Value

If spatial = FALSE a named vector of length four with the extent of the bounding box. Otherwise a SpatialPolygon or a simple feature polygon with the bounding box.

Examples

data(deer)
bbox(deer)
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 4309723 ymin: 3442716 xmax: 4318383 ymax: 3450713
#> Projected CRS: ETRS89 / LAEA Europe
#> POLYGON ((4309723 3442716, 4309723 3450713, 431...
bbox(deer, spatial = FALSE)
#>    xmin    ymin    xmax    ymax 
#> 4309723 3442716 4318383 3450713 
bbox(deer, buffer = 100, spatial = FALSE)
#>    xmin    ymin    xmax    ymax 
#> 4309623 3442616 4318483 3450813 

# For steps
deer |> steps_by_burst() |> bbox(spatial = FALSE)
#>    xmin    ymin    xmax    ymax 
#> 4309723 3442716 4318383 3450713 
deer |> steps_by_burst() |> bbox(buffer = 100, spatial = FALSE)
#>    xmin    ymin    xmax    ymax 
#> 4309623 3442616 4318483 3450813 
deer |> steps_by_burst() |> random_steps() |> bbox(spatial = FALSE)
#> Warning: Some bursts contain < 3 steps and will be removed
#>    xmin    ymin    xmax    ymax 
#> 4308779 3442277 4320449 3452008 

# Further manipulations are possible
deer |> bbox() |> sf::st_transform(4326)
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 9.827386 ymin: 54.09117 xmax: 9.960009 ymax: 54.16314
#> Geodetic CRS:  WGS 84
#> POLYGON ((9.827682 54.09117, 9.827386 54.16303,...