Extract covariate values

Extract the covariate values at relocations, or at the beginning or end of steps.

extract_covariates(x, ...)

# S3 method for track_xy
extract_covariates(x, covariates, ...)

# S3 method for random_points
extract_covariates(x, covariates, ...)

# S3 method for steps_xy
extract_covariates(x, covariates, where = "end", ...)

extract_covariates_along(x, ...)

# S3 method for steps_xy
extract_covariates_along(x, covariates, ...)

extract_covariates_var_time(x, ...)

# S3 method for track_xyt
extract_covariates_var_time(
  x,
  covariates,
  when = "any",
  max_time,
  name_covar = "time_var_covar",
  ...
)

# S3 method for steps_xyt
extract_covariates_var_time(
  x,
  covariates,
  when = "any",
  max_time,
  name_covar = "time_var_covar",
  where = "end",
  ...
)

Arguments

x

[track_xy, track_xyt, steps]
Either a track created with mk_track or track, or steps.

...

Additional arguments passed to terra::extract().

covariates

[SpatRaster]
The (environmental) covariates. For extract_covariates_var_time the argument covariates need to have a z-column (i.e. the time stamp).

where

[character(1)="end"]{"start", "end", "both"}
For steps this determines if the covariate values should be extracted at the beginning or the end of a step. or end.

when

[character(1)="any"]{"any", "before", "after"}
Specifies for for extract_covariates_var_time whether to look before, after or in both direction (any) for the temporally closest environmental raster.

max_time

[Period(1)]
The maximum time difference between a relocation and the corresponding raster. If no rasters are within the specified max. distance NA is returned.

name_covar

[character(1)="time_var_covar"]
The name of the new column.

Value

A tibble with additional columns for covariate values.

Details

extract_covariates_along extracts the covariates along a straight line between the start and the end point of a (random) step. It returns a list, which in most cases will have to be processed further.

Examples

data(deer)
sh_forest <- get_sh_forest()
mini_deer <- deer[1:20, ]
mini_deer |> extract_covariates(sh_forest)
#> # A tibble: 20 × 5
#>          x_       y_ t_                  burst_ forest
#>       <dbl>    <dbl> <dttm>               <dbl>  <dbl>
#>  1 4314068. 3445807. 2008-03-30 00:01:47      1      0
#>  2 4314053. 3445768. 2008-03-30 06:00:54      1      0
#>  3 4314105. 3445859. 2008-03-30 12:01:47      1      0
#>  4 4314044. 3445785. 2008-03-30 18:01:24      1      0
#>  5 4313015. 3445858. 2008-03-31 00:01:23      1      1
#>  6 4312860. 3445857. 2008-03-31 06:01:45      1      1
#>  7 4312854. 3445856. 2008-03-31 12:01:11      1      1
#>  8 4312858. 3445858. 2008-03-31 18:01:55      1      1
#>  9 4312745. 3445862. 2008-04-01 00:01:24      1      1
#> 10 4312651. 3446024. 2008-04-01 06:00:54      1      1
#> 11 4312649. 3446030. 2008-04-01 12:01:43      1      1
#> 12 4312619. 3446100. 2008-04-01 18:01:42      1      1
#> 13 4312663. 3446384. 2008-04-02 06:00:47      2      1
#> 14 4312665. 3446391. 2008-04-02 12:02:21      2      1
#> 15 4312665. 3446476. 2008-04-02 18:01:47      2      0
#> 16 4312010. 3446648. 2008-04-03 00:00:47      2      0
#> 17 4311612. 3446447. 2008-04-03 06:01:23      2      0
#> 18 4311617. 3446461. 2008-04-03 12:02:11      2      0
#> 19 4311593. 3446419. 2008-04-03 18:01:42      2      0
#> 20 4311570. 3446398. 2008-04-04 00:01:54      2      0
mini_deer |> steps() |> extract_covariates(sh_forest)
#> Warning: burst's are ignored, use steps_by_burst instead.
#> # A tibble: 19 × 11
#>         x1_      x2_      y1_     y2_    sl_ direc…¹     ta_ t1_                
#>       <dbl>    <dbl>    <dbl>   <dbl>  <dbl>   <dbl>   <dbl> <dttm>             
#>  1 4314068. 4314053. 3445807.  3.45e6 4.18e1  -1.95  NA      2008-03-30 00:01:47
#>  2 4314053. 4314105. 3445768.  3.45e6 1.04e2   1.05   3.00   2008-03-30 06:00:54
#>  3 4314105. 4314044. 3445859.  3.45e6 9.53e1  -2.27   2.97   2008-03-30 12:01:47
#>  4 4314044. 4313015. 3445785.  3.45e6 1.03e3   3.07  -0.945  2008-03-30 18:01:24
#>  5 4313015. 4312860. 3445858.  3.45e6 1.55e2  -3.14   0.0756 2008-03-31 00:01:23
#>  6 4312860. 4312854. 3445857.  3.45e6 6.08e0  -2.99   0.146  2008-03-31 06:01:45
#>  7 4312854. 4312858. 3445856.  3.45e6 4.47e0   0.449 -2.84   2008-03-31 12:01:11
#>  8 4312858. 4312745. 3445858.  3.45e6 1.13e2   3.11   2.66   2008-03-31 18:01:55
#>  9 4312745. 4312651. 3445862.  3.45e6 1.87e2   2.09  -1.02   2008-04-01 00:01:24
#> 10 4312651. 4312649. 3446024.  3.45e6 6.32e0   1.88  -0.216  2008-04-01 06:00:54
#> 11 4312649. 4312619. 3446030.  3.45e6 7.56e1   1.98   0.101  2008-04-01 12:01:43
#> 12 4312619. 4312663. 3446100.  3.45e6 2.88e2   1.42  -0.562  2008-04-01 18:01:42
#> 13 4312663. 4312665. 3446384.  3.45e6 7.28e0   1.28  -0.139  2008-04-02 06:00:47
#> 14 4312665. 4312665. 3446391.  3.45e6 8.50e1   1.57   0.290  2008-04-02 12:02:21
#> 15 4312665. 4312010. 3446476.  3.45e6 6.78e2   2.88   1.32   2008-04-02 18:01:47
#> 16 4312010. 4311612. 3446648.  3.45e6 4.46e2  -2.67   0.725  2008-04-03 00:00:47
#> 17 4311612. 4311617. 3446447.  3.45e6 1.49e1   1.21  -2.40   2008-04-03 06:01:23
#> 18 4311617. 4311593. 3446461.  3.45e6 4.79e1  -2.09   2.98   2008-04-03 12:02:11
#> 19 4311593. 4311570. 3446419.  3.45e6 3.11e1  -2.42  -0.330  2008-04-03 18:01:42
#> # … with 3 more variables: t2_ <dttm>, dt_ <drtn>, forest <dbl>, and
#> #   abbreviated variable name ¹​direction_p
mini_deer |> steps() |> extract_covariates(sh_forest, where = "start")
#> Warning: burst's are ignored, use steps_by_burst instead.
#> # A tibble: 19 × 11
#>         x1_      x2_      y1_     y2_    sl_ direc…¹     ta_ t1_                
#>       <dbl>    <dbl>    <dbl>   <dbl>  <dbl>   <dbl>   <dbl> <dttm>             
#>  1 4314068. 4314053. 3445807.  3.45e6 4.18e1  -1.95  NA      2008-03-30 00:01:47
#>  2 4314053. 4314105. 3445768.  3.45e6 1.04e2   1.05   3.00   2008-03-30 06:00:54
#>  3 4314105. 4314044. 3445859.  3.45e6 9.53e1  -2.27   2.97   2008-03-30 12:01:47
#>  4 4314044. 4313015. 3445785.  3.45e6 1.03e3   3.07  -0.945  2008-03-30 18:01:24
#>  5 4313015. 4312860. 3445858.  3.45e6 1.55e2  -3.14   0.0756 2008-03-31 00:01:23
#>  6 4312860. 4312854. 3445857.  3.45e6 6.08e0  -2.99   0.146  2008-03-31 06:01:45
#>  7 4312854. 4312858. 3445856.  3.45e6 4.47e0   0.449 -2.84   2008-03-31 12:01:11
#>  8 4312858. 4312745. 3445858.  3.45e6 1.13e2   3.11   2.66   2008-03-31 18:01:55
#>  9 4312745. 4312651. 3445862.  3.45e6 1.87e2   2.09  -1.02   2008-04-01 00:01:24
#> 10 4312651. 4312649. 3446024.  3.45e6 6.32e0   1.88  -0.216  2008-04-01 06:00:54
#> 11 4312649. 4312619. 3446030.  3.45e6 7.56e1   1.98   0.101  2008-04-01 12:01:43
#> 12 4312619. 4312663. 3446100.  3.45e6 2.88e2   1.42  -0.562  2008-04-01 18:01:42
#> 13 4312663. 4312665. 3446384.  3.45e6 7.28e0   1.28  -0.139  2008-04-02 06:00:47
#> 14 4312665. 4312665. 3446391.  3.45e6 8.50e1   1.57   0.290  2008-04-02 12:02:21
#> 15 4312665. 4312010. 3446476.  3.45e6 6.78e2   2.88   1.32   2008-04-02 18:01:47
#> 16 4312010. 4311612. 3446648.  3.45e6 4.46e2  -2.67   0.725  2008-04-03 00:00:47
#> 17 4311612. 4311617. 3446447.  3.45e6 1.49e1   1.21  -2.40   2008-04-03 06:01:23
#> 18 4311617. 4311593. 3446461.  3.45e6 4.79e1  -2.09   2.98   2008-04-03 12:02:11
#> 19 4311593. 4311570. 3446419.  3.45e6 3.11e1  -2.42  -0.330  2008-04-03 18:01:42
#> # … with 3 more variables: t2_ <dttm>, dt_ <drtn>, forest <dbl>, and
#> #   abbreviated variable name ¹​direction_p

# \donttest{
# Illustration of extracting covariates along the a step
mini_deer |> steps() |> random_steps() |>
  extract_covariates(sh_forest) |> # extract at the endpoint
  (\(.) mutate(., for_path = extract_covariates_along(., sh_forest)))()  |>
  # 1 = forest, lets calc the fraction of forest along the path
  mutate(for_per = purrr::map_dbl(for_path, function(x) mean(x == 1)))
#> Warning: burst's are ignored, use steps_by_burst instead.
#> # A tibble: 198 × 14
#>         x1_      x2_      y1_      y2_      sl_    ta_ t1_                
#>  *    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>  <dbl> <dttm>             
#>  1 4314053. 4314105. 3445768. 3445859.  104.     3.00  2008-03-30 06:00:54
#>  2 4314053. 4314091. 3445768. 3445706.   73.0    0.931 2008-03-30 06:00:54
#>  3 4314053. 4314051. 3445768. 3445773.    5.15  -2.39  2008-03-30 06:00:54
#>  4 4314053. 4313947. 3445768. 3445583.  213.    -0.138 2008-03-30 06:00:54
#>  5 4314053. 4314034. 3445768. 3445802.   38.4   -2.26  2008-03-30 06:00:54
#>  6 4314053. 4313966. 3445768. 3445655.  143.    -0.274 2008-03-30 06:00:54
#>  7 4314053. 4314053. 3445768. 3445768.    0.653  0.154 2008-03-30 06:00:54
#>  8 4314053. 4314054. 3445768. 3445768.    1.30   1.90  2008-03-30 06:00:54
#>  9 4314053. 4314052. 3445768. 3445770.    2.09  -2.65  2008-03-30 06:00:54
#> 10 4314053. 4315141. 3445768. 3444674. 1544.     1.16  2008-03-30 06:00:54
#> # … with 188 more rows, and 7 more variables: t2_ <dttm>, dt_ <drtn>,
#> #   case_ <lgl>, step_id_ <dbl>, forest <dbl>, for_path <list>, for_per <dbl>
# }