Least square cross validation bandwidth

Use least square cross validation (lscv) to estimate bandwidth for kernel home-range estimation.

hr_kde_lscv(
  x,
  range = do.call(seq, as.list(c(hr_kde_ref(x) * c(0.1, 2), length.out = 100))),
  which_min = "global",
  rescale = "none",
  trast = make_trast(x)
)

Arguments

x

[track_xy, track_xyt]
A track created with make_track.

range

numeric vector with different candidate h values.

which_min

A character indicating if the global or local minimum should be searched for.

rescale

[character(1)]
Rescaling method for reference bandwidth calculation. Must be one of "unitvar", "xvar", or "none".

trast

A template raster.

Value

vector of length two.

Details

hr_kde_lscv calculates least square cross validation bandwidth. This implementation is based on Seaman and Powell (1996). If whichMin is "global" the global minimum is returned, else the local minimum with the largest candidate bandwidth is returned.

References

Seaman, D. E., & Powell, R. A. (1996). An evaluation of the accuracy of kernel density estimators for home range analysis. Ecology, 77(7), 2075-2085.