Home ranges

Functions to calculate animal home ranges from a track_xy*. hr_mcp, hr_kde, and hr_locoh calculate the minimum convex polygon, kernel density, and local convex hull home range respectively.

hr_akde(x, ...)

# S3 method for track_xyt
hr_akde(
  x,
  model = fit_ctmm(x, "iid"),
  keep.data = TRUE,
  trast = make_trast(x),
  levels = 0.95,
  wrap = FALSE,
  ...
)

hr_kde(x, ...)

# S3 method for track_xy
hr_kde(
  x,
  h = hr_kde_ref(x),
  trast = make_trast(x),
  levels = 0.95,
  keep.data = TRUE,
  wrap = FALSE,
  ...
)

hr_locoh(x, ...)

# S3 method for track_xy
hr_locoh(
  x,
  n = 10,
  type = "k",
  levels = 0.95,
  keep.data = TRUE,
  rand_buffer = 1e-05,
  ...
)

hr_mcp(x, ...)

hr_od(x, ...)

Arguments

x

[track_xy, track_xyt]
A track created with make_track.

...

Further arguments, none implemented.

model

A continuous time movement model. This can be fitted either with ctmm::ctmm.fit or fit_ctmm.

keep.data

[logic(1)]
Should the original tracking data be included in the estimate?

trast

[SpatRast]
A template raster for kernel density home-ranges.

levels

[numeric]
The isopleth levels used for calculating home ranges. Should be 0 < level < 1.

wrap

[logical(1)]
If TRUE the UD is wrapped (see terra::wrap()).

h

[numeric(2)]
The bandwidth for kernel density estimation.

n

[integer(1)]
The number of neighbors used when calculating local convex hulls.

type

k, r or a. Type of LoCoH.

rand_buffer

[numeric(1)]
Random buffer to avoid polygons with area 0 (if coordinates are numerically identical).

Value

A hr-estimate.

References

Worton, B. J. (1989). Kernel methods for estimating the utilization distribution in home-range studies. Ecology, 70(1), 164-168. C. H. Fleming, W. F. Fagan, T. Mueller, K. A. Olson, P. Leimgruber, J. M. Calabrese, “Rigorous home-range estimation with movement data: A new autocorrelated kernel-density estimator”, Ecology, 96:5, 1182-1188 (2015).

Fleming, C. H., Fagan, W. F., Mueller, T., Olson, K. A., Leimgruber, P., & Calabrese, J. M. (2016). Estimating where and how animals travel: an optimal framework for path reconstruction from autocorrelated tracking data. Ecology, 97(3), 576-582.

Examples

# \donttest{
data(deer)
mini_deer <- deer[1:100, ]

# MCP ---------------------------------------------------------------------
mcp1 <- hr_mcp(mini_deer)
hr_area(mcp1)
#> # A tibble: 1 × 3
#>   level what         area
#>   <dbl> <chr>       <dbl>
#> 1  0.95 estimate 2226718.

# calculated MCP at different levels
mcp1 <- hr_mcp(mini_deer, levels = seq(0.3, 1, 0.1))
hr_area(mcp1)
#> # A tibble: 8 × 3
#>   level what         area
#>   <dbl> <chr>       <dbl>
#> 1   1   estimate 3827291.
#> 2   0.9 estimate 2223269.
#> 3   0.8 estimate 1464224.
#> 4   0.7 estimate 1437598.
#> 5   0.6 estimate  767768.
#> 6   0.5 estimate  732019.
#> 7   0.4 estimate  729211.
#> 8   0.3 estimate  639048.

# CRS are inherited
get_crs(mini_deer)
#> Coordinate Reference System:
#>   User input: EPSG:3035 
#>   wkt:
#> PROJCS["ETRS89 / LAEA Europe",
#>     GEOGCS["ETRS89",
#>         DATUM["European_Terrestrial_Reference_System_1989",
#>             SPHEROID["GRS 1980",6378137,298.257222101,
#>                 AUTHORITY["EPSG","7019"]],
#>             TOWGS84[0,0,0,0,0,0,0],
#>             AUTHORITY["EPSG","6258"]],
#>         PRIMEM["Greenwich",0,
#>             AUTHORITY["EPSG","8901"]],
#>         UNIT["degree",0.0174532925199433,
#>             AUTHORITY["EPSG","9122"]],
#>         AUTHORITY["EPSG","4258"]],
#>     PROJECTION["Lambert_Azimuthal_Equal_Area"],
#>     PARAMETER["latitude_of_center",52],
#>     PARAMETER["longitude_of_center",10],
#>     PARAMETER["false_easting",4321000],
#>     PARAMETER["false_northing",3210000],
#>     UNIT["metre",1,
#>         AUTHORITY["EPSG","9001"]],
#>     AUTHORITY["EPSG","3035"]]
mcps <- hr_mcp(mini_deer, levels = c(0.5, 0.95, 1))
has_crs(mcps)
#> [1] TRUE

# Kernel density estimaiton (KDE) -----------------------------------------
kde1 <- hr_kde(mini_deer)
hr_area(kde1)
#> # A tibble: 1 × 3
#>   level what         area
#>   <dbl> <chr>       <dbl>
#> 1  0.95 estimate 9622684.
get_crs(kde1)
#> Coordinate Reference System:
#>   User input: EPSG:3035 
#>   wkt:
#> PROJCS["ETRS89 / LAEA Europe",
#>     GEOGCS["ETRS89",
#>         DATUM["European_Terrestrial_Reference_System_1989",
#>             SPHEROID["GRS 1980",6378137,298.257222101,
#>                 AUTHORITY["EPSG","7019"]],
#>             TOWGS84[0,0,0,0,0,0,0],
#>             AUTHORITY["EPSG","6258"]],
#>         PRIMEM["Greenwich",0,
#>             AUTHORITY["EPSG","8901"]],
#>         UNIT["degree",0.0174532925199433,
#>             AUTHORITY["EPSG","9122"]],
#>         AUTHORITY["EPSG","4258"]],
#>     PROJECTION["Lambert_Azimuthal_Equal_Area"],
#>     PARAMETER["latitude_of_center",52],
#>     PARAMETER["longitude_of_center",10],
#>     PARAMETER["false_easting",4321000],
#>     PARAMETER["false_northing",3210000],
#>     UNIT["metre",1,
#>         AUTHORITY["EPSG","9001"]],
#>     AUTHORITY["EPSG","3035"]]

# akde
data(deer)
mini_deer <- deer[1:20, ]
ud1 <- hr_akde(mini_deer) # uses an iid ctmm
ud2 <- hr_akde(mini_deer, model = fit_ctmm(deer, "ou")) # uses an OU ctmm
# }
# od
# \donttest{
data(deer)
ud1 <- hr_od(deer) # uses an iid ctmm
ud2 <- hr_akde(deer, model = fit_ctmm(deer, "ou")) # uses an OU ctmm
# }