Create an initial step for simulations

An initial step for simulations. This step can either be created by defining a step from scratch or by using an observed step.

make_start(x, ...)

# S3 method for numeric
make_start(
  x = c(0, 0),
  ta_ = 0,
  time = Sys.time(),
  dt = hours(1),
  crs = NA,
  ...
)

# S3 method for track_xyt
make_start(x, ta_ = 0, dt = hours(1), ...)

# S3 method for steps_xyt
make_start(x, ...)

Arguments

x

[steps_xyt,numeric(2)]
A step of class steps_xyt or the start coordinates..

...

Further arguments, none implemented.

ta_

[numeric(1)]{0}
The initial turn-angle.

time

[POSIXt(1)]{Sys.time()}
The time stamp when the simulation starts.

dt

[Period(1)]{hours(1)}
The sampling rate of the simulations.

crs

[int(1)]{NA}
The coordinate reference system of the start location given as EPSG code.