Create a redistribution kernel

From a fitted integrated step-selection function for a given position a redistribution kernel is calculated (i.e., the product of the movement kernel and the selection function).

redistribution_kernel(
  x = make_issf_model(),
  start = make_start(),
  map,
  fun = function(xy, map) {
     extract_covariates(xy, map, where = "both")
 },
  covars = NULL,
  max.dist = get_max_dist(x),
  n.control = 1e+06,
  n.sample = 1,
  stochastic = FALSE,
  compensate.movement = !stochastic,
  normalize = TRUE,
  interpolate = FALSE,
  as.rast = FALSE,
  tolerance.outside = 0
)

Arguments

x

[fit_issf]
A fitted integrated step-selection function. Generated either with fit_issf() or make make_issf_model().

start

[sim_start]
The start position in space and time. See make_start().

map

[SpatRaster]
A SpatRaster with all covariates.

fun

[function]
A function that is executed on each location of the redistribution kernel. The default function is extract_covariates().

covars

[tibble]
Additional covariates that might be used in the model (e.g., time of day).

max.dist

[numeric(1)]
The maximum distance of the redistribution kernel.

n.control

[integer(1)]{1e6}
The number of points of the redistribution kernel (this is only important if stochastic = TRUE).

n.sample

[integer(1)]{1}
The number of points sampled from the redistribution kernel (this is only important if as.rast = FALSE).

stochastic

[logical(1)]{FALSE}
If TRUE the redistribution kernel is sampled using a random sample of size n.control. Otherwise each cell in the redistribution kernel is used.

compensate.movement

[logical(1)]
Indicates if movement parameters are corrected or not. This only relevant if stochastic = FALSE.

normalize

[logical(1)]{TRUE}
If TRUE the redistribution kernel is normalized to sum to one.

interpolate

[logical(1)]{FALSE}
If TRUE a stochastic redistribution kernel is interpolated to return a raster layer. Note, this is just for completeness and is computationally inefficient in most situations.

as.rast

[logical(1)]{TRUE}
If TRUE a SpatRaster should be returned.

tolerance.outside

[numeric(1)]{0}
The proportion of the redistribution kernel that is allowed to be outside the map.