Returns a summary of sampling rates

Returns a summary of sampling rates

summarize_sampling_rate(x, ...)

# S3 method for track_xyt
summarize_sampling_rate(
  x,
  time_unit = "auto",
  summarize = TRUE,
  as_tibble = TRUE,
  ...
)

summarize_sampling_rate_many(x, ...)

# S3 method for track_xyt
summarize_sampling_rate_many(x, cols, time_unit = "auto", ...)

Arguments

x

A track_xyt.

...

Further arguments, none implemented.

time_unit

[character(1) = "auto"]
Which time unit will be used.

summarize

A logical. If TRUE a summary is returned, otherwise raw sampling intervals are returned.

as_tibble

A logical. Should result be returned as tibble or as table.

cols

Columns used for grouping.

Value

Depending on summarize and as_tibble, a vector, table or tibble.

Examples

data(deer)
amt::summarize_sampling_rate(deer)
#> # A tibble: 1 × 9
#>     min    q1 median  mean    q3   max    sd     n unit 
#>   <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <int> <chr>
#> 1  5.96  6.00   6.01  11.5  6.02 3924.  137.   825 hour 

data(amt_fisher)
# Add the month
amt_fisher |> mutate(yday = lubridate::yday(t_)) |>
summarize_sampling_rate_many(c("id", "yday"))
#> # A tibble: 114 × 11
#>    id     yday   min    q1 median  mean    q3   max      sd     n unit 
#>  * <chr> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>   <dbl> <int> <chr>
#>  1 M1       42  14.2  14.7   14.9  15.0  15.1  16.0   0.381    46 min  
#>  2 M1       43  14.6  14.8   15.0  32.1  15.3 390.   79.9      22 min  
#>  3 M1       44  14.4  14.7   15.0  16.5  15.4  45.6   6.17     50 min  
#>  4 M1       45  13.8  14.7   15.1  23.4  15.7 179.   30.7      34 min  
#>  5 M1       46  13.4  14.8   15.0  29.4  15.4 631.   88.9      48 min  
#>  6 M1       47  13.7  14.7   15.0  15.7  15.3  30.5   3.12     89 min  
#>  7 M1       48  14.0  14.7   15.1  33.9  15.6 691.  104.       42 min  
#>  8 M1       49  13.7  14.7   15.0  37.5  15.4 781.  124.       38 min  
#>  9 M1       50  14.1  14.8   15.3  31.9  33.7 120.   31.5      16 min  
#> 10 M1       51  13.3  14.5   15.1  52.7  15.6 901.  171.       27 min  
#> # ℹ 104 more rows