Time of the day when a fix was taken

A convenience wrapper around suncalc::getSunlightTimes to annotate if a fix was taken during day or night (optionally also include dawn and dusk).

time_of_day(x, ...)

# S3 method for track_xyt
time_of_day(x, include.crepuscule = FALSE, ...)

# S3 method for steps_xyt
time_of_day(x, include.crepuscule = FALSE, where = "end", ...)

Arguments

x

[track_xyt,steps_xyt]
A track or steps.

...

Further arguments, none implemented.

include.crepuscule

[logical(1)=TRUE]
Should dawn and dusk be included.

where

[character(1)="end"]{"start", "end", "both"} For steps, should the start, end or both time points be used?

Value

A tibble with an additional column tod_ that contains the time of the day for each relocation.

Examples

data(deer)
deer |> time_of_day()
#> # A tibble: 826 × 5
#>          x_       y_ t_                         burst_ tod_ 
#>       <dbl>    <dbl> <dttm>                      <dbl> <fct>
#>  1 4314068. 3445807. 2008-03-30 00:01:47.000000      1 night
#>  2 4314053. 3445768. 2008-03-30 06:00:54.000000      1 day  
#>  3 4314105. 3445859. 2008-03-30 12:01:47.000000      1 day  
#>  4 4314044. 3445785. 2008-03-30 18:01:24.000000      1 night
#>  5 4313015. 3445858. 2008-03-31 00:01:23.000000      1 night
#>  6 4312860. 3445857. 2008-03-31 06:01:45.000000      1 day  
#>  7 4312854. 3445856. 2008-03-31 12:01:11.000000      1 day  
#>  8 4312858. 3445858. 2008-03-31 18:01:55.000000      1 night
#>  9 4312745. 3445862. 2008-04-01 00:01:24.000000      1 night
#> 10 4312651. 3446024. 2008-04-01 06:00:54.000000      1 day  
#> # ℹ 816 more rows
deer |> steps_by_burst() |> time_of_day()
#> # A tibble: 791 × 12
#>    burst_      x1_      x2_      y1_      y2_     sl_ direction_p     ta_
#>     <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>       <dbl>   <dbl>
#>  1      1 4314068. 4314053. 3445807. 3445768.   41.8       -1.95  NA     
#>  2      1 4314053. 4314105. 3445768. 3445859.  104.         1.05   3.00  
#>  3      1 4314105. 4314044. 3445859. 3445785.   95.3       -2.27   2.97  
#>  4      1 4314044. 4313015. 3445785. 3445858. 1032.         3.07  -0.945 
#>  5      1 4313015. 4312860. 3445858. 3445857.  155.        -3.14   0.0756
#>  6      1 4312860. 4312854. 3445857. 3445856.    6.08      -2.99   0.146 
#>  7      1 4312854. 4312858. 3445856. 3445858.    4.47       0.449 -2.84  
#>  8      1 4312858. 4312745. 3445858. 3445862.  113.         3.11   2.66  
#>  9      1 4312745. 4312651. 3445862. 3446024.  187.         2.09  -1.02  
#> 10      1 4312651. 4312649. 3446024. 3446030.    6.32       1.88  -0.216 
#> # ℹ 781 more rows
#> # ℹ 4 more variables: t1_ <dttm>, t2_ <dttm>, dt_ <drtn>, tod_end_ <fct>
deer |> steps_by_burst() |> time_of_day(where = "start")
#> # A tibble: 791 × 12
#>    burst_      x1_      x2_      y1_      y2_     sl_ direction_p     ta_
#>     <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>       <dbl>   <dbl>
#>  1      1 4314068. 4314053. 3445807. 3445768.   41.8       -1.95  NA     
#>  2      1 4314053. 4314105. 3445768. 3445859.  104.         1.05   3.00  
#>  3      1 4314105. 4314044. 3445859. 3445785.   95.3       -2.27   2.97  
#>  4      1 4314044. 4313015. 3445785. 3445858. 1032.         3.07  -0.945 
#>  5      1 4313015. 4312860. 3445858. 3445857.  155.        -3.14   0.0756
#>  6      1 4312860. 4312854. 3445857. 3445856.    6.08      -2.99   0.146 
#>  7      1 4312854. 4312858. 3445856. 3445858.    4.47       0.449 -2.84  
#>  8      1 4312858. 4312745. 3445858. 3445862.  113.         3.11   2.66  
#>  9      1 4312745. 4312651. 3445862. 3446024.  187.         2.09  -1.02  
#> 10      1 4312651. 4312649. 3446024. 3446030.    6.32       1.88  -0.216 
#> # ℹ 781 more rows
#> # ℹ 4 more variables: t1_ <dttm>, t2_ <dttm>, dt_ <drtn>, tod_start_ <fct>
deer |> steps_by_burst() |> time_of_day(where = "end")
#> # A tibble: 791 × 12
#>    burst_      x1_      x2_      y1_      y2_     sl_ direction_p     ta_
#>     <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>       <dbl>   <dbl>
#>  1      1 4314068. 4314053. 3445807. 3445768.   41.8       -1.95  NA     
#>  2      1 4314053. 4314105. 3445768. 3445859.  104.         1.05   3.00  
#>  3      1 4314105. 4314044. 3445859. 3445785.   95.3       -2.27   2.97  
#>  4      1 4314044. 4313015. 3445785. 3445858. 1032.         3.07  -0.945 
#>  5      1 4313015. 4312860. 3445858. 3445857.  155.        -3.14   0.0756
#>  6      1 4312860. 4312854. 3445857. 3445856.    6.08      -2.99   0.146 
#>  7      1 4312854. 4312858. 3445856. 3445858.    4.47       0.449 -2.84  
#>  8      1 4312858. 4312745. 3445858. 3445862.  113.         3.11   2.66  
#>  9      1 4312745. 4312651. 3445862. 3446024.  187.         2.09  -1.02  
#> 10      1 4312651. 4312649. 3446024. 3446030.    6.32       1.88  -0.216 
#> # ℹ 781 more rows
#> # ℹ 4 more variables: t1_ <dttm>, t2_ <dttm>, dt_ <drtn>, tod_end_ <fct>
deer |> steps_by_burst() |> time_of_day(where = "both")
#> # A tibble: 791 × 13
#>    burst_      x1_      x2_      y1_      y2_     sl_ direction_p     ta_
#>     <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>       <dbl>   <dbl>
#>  1      1 4314068. 4314053. 3445807. 3445768.   41.8       -1.95  NA     
#>  2      1 4314053. 4314105. 3445768. 3445859.  104.         1.05   3.00  
#>  3      1 4314105. 4314044. 3445859. 3445785.   95.3       -2.27   2.97  
#>  4      1 4314044. 4313015. 3445785. 3445858. 1032.         3.07  -0.945 
#>  5      1 4313015. 4312860. 3445858. 3445857.  155.        -3.14   0.0756
#>  6      1 4312860. 4312854. 3445857. 3445856.    6.08      -2.99   0.146 
#>  7      1 4312854. 4312858. 3445856. 3445858.    4.47       0.449 -2.84  
#>  8      1 4312858. 4312745. 3445858. 3445862.  113.         3.11   2.66  
#>  9      1 4312745. 4312651. 3445862. 3446024.  187.         2.09  -1.02  
#> 10      1 4312651. 4312649. 3446024. 3446030.    6.32       1.88  -0.216 
#> # ℹ 781 more rows
#> # ℹ 5 more variables: t1_ <dttm>, t2_ <dttm>, dt_ <drtn>, tod_start_ <fct>,
#> #   tod_end_ <fct>