Update movement distributions

Update tentative step length or turning angle distribution from a fitted iSSF.

update_sl_distr(
  object,
  beta_sl = "sl_",
  beta_log_sl = "log_sl_",
  beta_sl_sq = "sl_sq_",
  beta_log_sl_sq = "log_sl_sq_",
  ...
)

update_ta_distr(object, beta_cos_ta = "cos_ta_", ...)

Arguments

object

[fit_clogit]
A fitted iSSF model.

beta_sl

[character]
The name of the coefficient of the step length.

beta_log_sl

[character]
The name of the coefficient of the log of the step length.

beta_sl_sq

[character]
The name of the coefficient of the square of the step length.

beta_log_sl_sq

[character]
The name of the coefficient of the square of log of the step length.

...

Further arguments, none implemented.

beta_cos_ta

[character]
The name of the coefficient of cosine of the turning angle.

Value

An amt_distr object, which consists of a list with the name of the distribution and its parameters (saved in params).

References

Fieberg J, Signer J, Smith BJ, Avgar T (2020). “A “How-to” Guide for Interpreting Parameters in Resource-and Step-Selection Analyses.” bioRxiv.

See also

Author

Brian J. Smith and Johannes Signer

Examples

# Fit an SSF, then update movement parameters.
data(deer)
mini_deer <- deer[1:100, ]
sh_forest <- get_sh_forest()

# Prepare data for SSF
ssf_data <- mini_deer |>
  steps_by_burst() |>
  random_steps(n = 15) |>
  extract_covariates(sh_forest) |>
  mutate(forest = factor(forest, levels = 1:0,
                    labels = c("forest", "non-forest")),
  cos_ta_ = cos(ta_),
  log_sl_ = log(sl_))
#> Warning: Some bursts contain < 3 steps and will be removed

# Check tentative distributions
# Step length
sl_distr_params(ssf_data)
#> $shape
#> [1] 0.5799379
#> 
#> $scale
#> [1] 343.194
#> 
attr(ssf_data, "sl_")
#> $name
#> [1] "gamma"
#> 
#> $params
#> $params$shape
#> [1] 0.5799379
#> 
#> $params$scale
#> [1] 343.194
#> 
#> 
#> $vcov
#>             scale        shape
#> scale 3874.346331 -2.935218369
#> shape   -2.935218  0.004960626
#> 
#> attr(,"class")
#> [1] "gamma_distr" "sl_distr"    "amt_distr"   "list"       
#    Turning angle
ta_distr_params(ssf_data)
#> $kappa
#> [1] 0.2825356
#> 
#> $mu
#> Circular Data: 
#> Type = angles 
#> Units = radians 
#> Template = none 
#> Modulo = asis 
#> Zero = 0 
#> Rotation = counter 
#> [1] 0
#> 

# Fit an iSSF
m1 <- ssf_data |>
  fit_issf(case_ ~ forest +
               sl_ + log_sl_ + cos_ta_ +
               strata(step_id_))

# Update step length distribution
new_gamma <- update_sl_distr(m1)

# Update turning angle distribution
new_vm <- update_ta_distr(m1)

# It is also possible to use different step length distributions

# exponential step-length distribution
s2 <- mini_deer |> steps_by_burst()
s2 <- random_steps(s2, sl_distr = fit_distr(s2$sl_, "exp"))
#> Warning: Some bursts contain < 3 steps and will be removed
m2 <- s2 |>
  fit_clogit(case_ ~ sl_ + strata(step_id_))
update_sl_distr(m2)
#> $name
#> [1] "exp"
#> 
#> $params
#> $params$rate
#> [1] 0.005255201
#> 
#> 
#> $vcov
#> NULL
#> 
#> attr(,"class")
#> [1] "exp_distr" "sl_distr"  "amt_distr" "list"     

# half normal step-length distribution
s3 <- mini_deer |> steps_by_burst()
s3 <- random_steps(s3, sl_distr = fit_distr(s3$sl_, "hnorm"))
#> Warning: Some bursts contain < 3 steps and will be removed
m3 <- s3 |>
  mutate(sl_sq_ = sl_^2) |>
  fit_clogit(case_ ~ sl_sq_ + strata(step_id_))
update_sl_distr(m3)
#> $name
#> [1] "hnorm"
#> 
#> $params
#> $params$sd
#> [1] 396.9396
#> 
#> 
#> $vcov
#> NULL
#> 
#> attr(,"class")
#> [1] "hnorm_distr" "sl_distr"    "amt_distr"   "list"       

# log normal step-length distribution
s4 <- mini_deer |> steps_by_burst()
s4 <- random_steps(s4, sl_distr = fit_distr(s4$sl_, "lnorm"))
#> Warning: Some bursts contain < 3 steps and will be removed
m4 <- s4 |>
  mutate(log_sl_ = log(sl_), log_sl_sq_ = log(sl_)^2) |>
  fit_clogit(case_ ~ log_sl_ + log_sl_sq_ + strata(step_id_))
update_sl_distr(m4)
#> $name
#> [1] "lnorm"
#> 
#> $params
#> $params$meanlog
#> [1] 3.712034
#> 
#> $params$sdlog
#> [1] 1.401068
#> 
#> 
#> $vcov
#> NULL
#> 
#> attr(,"class")
#> [1] "lnorm_distr" "sl_distr"    "amt_distr"   "list"