Inspect a track

Provides a very basic interface to leaflet and lets the user inspect relocations on an interactive map.

inspect(x, ...)

# S3 method for track_xy
inspect(x, popup = NULL, cluster = TRUE, ...)

Arguments

x

[track_xy, track_xyt]
A track created with make_track.

...

Further arguments, none implemented.

popup

[character(nrow(x))]
Optional labels for popups.

cluster

[logical(1)]
If TRUE points are clustered at lower zoom levels.

Value

An interactive leaflet map.

Note

Important, x requires a valid coordinate reference system.

See also

Examples

data(sh)
x <- track(x = sh$x, y = sh$y, crs = 31467)

# \donttest{
inspect(x)

inspect(x, cluster = FALSE)

inspect(x, popup = 1:nrow(x), cluster = FALSE)

# }